hv: vm: Move some common helpers to common scope
No logic change. Only movement.

Tracked-On: #8830
Acked-by: Wang Yu1 <yu1.wang@intel.com>
