+++ hypervisor/arch/riscv/guest/vm.c
@@ -10,22 +10,6 @@
 
 #include <asm/guest/vm.h>
 
-/* FIXME */
-struct acrn_vm *get_vm_from_vmid(__unused uint16_t vm_id)
-{
-	return NULL;
-}
-
-bool is_paused_vm(__unused const struct acrn_vm *vm)
-{
-	return false;
-}
-
-bool is_poweroff_vm(__unused const struct acrn_vm *vm)
-{
-	return true;
-}
-
 void arch_trigger_level_intr(__unused struct acrn_vm *vm,
 			__unused uint32_t irq, __unused bool assert) {}
