+++ hypervisor/common/vm.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+#include <per_cpu.h>
 #include <vcpu.h>
 #include <vm.h>
 #include <logmsg.h>
@@ -15,6 +16,166 @@ static struct acrn_vm vm_array[CONFIG_MAX_VM_NUM] __aligned(PAGE_SIZE);
 
 static struct acrn_vm *service_vm_ptr = NULL;
 
+uint16_t get_unused_vmid(void)
+{
+	uint16_t vm_id;
+	struct acrn_vm_config *vm_config;
+
+	for (vm_id = 0; vm_id < CONFIG_MAX_VM_NUM; vm_id++) {
+		vm_config = get_vm_config(vm_id);
+		if ((vm_config->name[0] == '\0') && ((vm_config->guest_flags & GUEST_FLAG_STATIC_VM) == 0U)) {
+			break;
+		}
+	}
+	return (vm_id < CONFIG_MAX_VM_NUM) ? (vm_id) : (ACRN_INVALID_VMID);
+}
+
+uint16_t get_vmid_by_name(const char *name)
+{
+	uint16_t vm_id;
+
+	for (vm_id = 0U; vm_id < CONFIG_MAX_VM_NUM; vm_id++) {
+		if ((*name != '\0') && vm_has_matched_name(vm_id, name)) {
+			break;
+		}
+	}
+	return (vm_id < CONFIG_MAX_VM_NUM) ? (vm_id) : (ACRN_INVALID_VMID);
+}
+
+/**
+ * @pre vm != NULL
+ */
+bool is_poweroff_vm(const struct acrn_vm *vm)
+{
+	return (vm->state == VM_POWERED_OFF);
+}
+
+/**
+ * @pre vm != NULL
+ */
+bool is_created_vm(const struct acrn_vm *vm)
+{
+	return (vm->state == VM_CREATED);
+}
+
+bool is_service_vm(const struct acrn_vm *vm)
+{
+	return (vm != NULL)  && (get_vm_config(vm->vm_id)->load_order == SERVICE_VM);
+}
+
+/**
+ * @pre vm != NULL
+ * @pre vm->vmid < CONFIG_MAX_VM_NUM
+ */
+bool is_postlaunched_vm(const struct acrn_vm *vm)
+{
+	return (get_vm_config(vm->vm_id)->load_order == POST_LAUNCHED_VM);
+}
+
+/**
+ * @pre vm != NULL
+ * @pre vm->vmid < CONFIG_MAX_VM_NUM
+ */
+bool is_prelaunched_vm(const struct acrn_vm *vm)
+{
+	struct acrn_vm_config *vm_config;
+
+	vm_config = get_vm_config(vm->vm_id);
+	return (vm_config->load_order == PRE_LAUNCHED_VM);
+}
+
+/**
+ * @pre vm != NULL && vm_config != NULL && vm->vmid < CONFIG_MAX_VM_NUM
+ */
+bool is_rt_vm(const struct acrn_vm *vm)
+{
+	struct acrn_vm_config *vm_config = get_vm_config(vm->vm_id);
+
+	return ((vm_config->guest_flags & GUEST_FLAG_RT) != 0U);
+}
+
+/**
+ * @pre vm != NULL && vm_config != NULL && vm->vmid < CONFIG_MAX_VM_NUM
+ *
+ * Stateful VM refers to VM that has its own state (such as internal file cache),
+ * and will experience state loss (file system corruption) if force powered down.
+ */
+bool is_stateful_vm(const struct acrn_vm *vm)
+{
+	struct acrn_vm_config *vm_config = get_vm_config(vm->vm_id);
+
+	/* TEE VM has GUEST_FLAG_STATELESS set implicitly */
+	return ((vm_config->guest_flags & GUEST_FLAG_STATELESS) == 0U);
+}
+
+/**
+ * @pre vm != NULL && vm_config != NULL && vm->vmid < CONFIG_MAX_VM_NUM
+ */
+bool is_static_configured_vm(const struct acrn_vm *vm)
+{
+	struct acrn_vm_config *vm_config = get_vm_config(vm->vm_id);
+
+	return ((vm_config->guest_flags & GUEST_FLAG_STATIC_VM) != 0U);
+}
+
+struct acrn_vm *get_highest_severity_vm(bool runtime)
+{
+	uint16_t vm_id, highest_vm_id = 0U;
+
+	for (vm_id = 1U; vm_id < CONFIG_MAX_VM_NUM; vm_id++) {
+		if (runtime && is_poweroff_vm(get_vm_from_vmid(vm_id))) {
+			/* If vm is non-existed or shutdown, it's not highest severity VM */
+			continue;
+		}
+
+		if (get_vm_severity(vm_id) > get_vm_severity(highest_vm_id)) {
+			highest_vm_id = vm_id;
+		}
+	}
+
+	return get_vm_from_vmid(highest_vm_id);
+}
+
+/**
+ * @pre vm != NULL
+ */
+void poweroff_if_rt_vm(struct acrn_vm *vm)
+{
+	if (is_rt_vm(vm) && !is_paused_vm(vm) && !is_poweroff_vm(vm)) {
+		vm->state = VM_READY_TO_POWEROFF;
+	}
+}
+
+/**
+ * if there is RT VM return true otherwise return false.
+ */
+bool has_rt_vm(void)
+{
+	uint16_t vm_id;
+
+	for (vm_id = 0U; vm_id < CONFIG_MAX_VM_NUM; vm_id++) {
+		if (is_rt_vm(get_vm_from_vmid(vm_id))) {
+			break;
+		}
+	}
+
+	return (vm_id != CONFIG_MAX_VM_NUM);
+}
+
+/*
+ * @pre vm != NULL
+ */
+void get_vm_lock(struct acrn_vm *vm)
+{
+	spinlock_obtain(&vm->vm_state_lock);
+}
+/*
+ * @pre vm != NULL
+ */
+void put_vm_lock(struct acrn_vm *vm)
+{
+	spinlock_release(&vm->vm_state_lock);
+}
 /**
  * @pre vm != NULL
  */
@@ -268,4 +429,37 @@ int32_t reset_vm(struct acrn_vm *vm)
 	}
 
 	return ret;
+}
+
+void make_shutdown_vm_request(uint16_t pcpu_id)
+{
+	bitmap_set(NEED_SHUTDOWN_VM, &per_cpu(pcpu_flag, pcpu_id));
+	if (get_pcpu_id() != pcpu_id) {
+		arch_smp_call_kick_pcpu(pcpu_id);
+	}
+}
+
+bool need_shutdown_vm(uint16_t pcpu_id)
+{
+	return bitmap_test_and_clear(NEED_SHUTDOWN_VM, &per_cpu(pcpu_flag, pcpu_id));
+}
+
+void shutdown_vm_from_idle(uint16_t pcpu_id)
+{
+	uint16_t vm_id;
+	uint64_t *vms = &per_cpu(shutdown_vm_bitmap, pcpu_id);
+	struct acrn_vm *vm;
+
+	for (vm_id = fls64(*vms); vm_id < CONFIG_MAX_VM_NUM; vm_id = fls64(*vms)) {
+		vm = get_vm_from_vmid(vm_id);
+		get_vm_lock(vm);
+		if (is_paused_vm(vm)) {
+			(void)destroy_vm(vm);
+			if (is_ready_for_system_shutdown()) {
+				shutdown_system();
+			}
+		}
+		put_vm_lock(vm);
+		bitmap_clear_non_atomic(vm_id, vms);
+	}
 }
