+++ hypervisor/include/arch/riscv/asm/guest/vm_reset.h
@@ -7,10 +7,4 @@
 #ifndef RISCV_VM_RESET_H_
 #define RISCV_VM_RESET_H_
 
-/* FIXME: riscv dummy function */
-static inline void shutdown_vm_from_idle(uint16_t pcpu_id)
-{
-	(void)pcpu_id;
-}
-
 #endif /* RISCV_VM_RESET_H_ */
