+++ hypervisor/include/arch/x86/asm/guest/vm_reset.h
@@ -10,7 +10,6 @@
 #include <acrn_common.h>
 
 void register_reset_port_handler(struct acrn_vm *vm);
-void shutdown_vm_from_idle(uint16_t pcpu_id);
 void triple_fault_shutdown_vm(struct acrn_vcpu *vcpu);
 
 #endif /* VM_RESET_H_ */
